Add simple function to get data from a CFURL source without using deprecated CFURLCreateDataAndPropertiesFromResource. More migration to SpatcialMixer from 3DMixer.

+++ Source/OpenAL/oalDevice.h
@@ -188,7 +188,7 @@ public:
             erase(it);
     }
 	
-    UInt32 Size () const { return size(); }
+    UInt32 Size () const { return UInt32(size()); }
     bool Empty () const { return empty(); }
 };
